% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvr.R
\name{broadcastPost}
\alias{broadcastPost}
\title{Broadcast to a server via POST - hence this can run in another process or even on another host.}
\usage{
broadcastPost(data, server = getOption("plotVR.broadcast.server", getURL()),
  ...)
}
\arguments{
\item{data}{the JSON-model data to broadcast, from \link{writeData}}

\item{server}{to which post to post the data, default is to \link{getURL} but also can be set
using \code{option(plotVR.broadcast.server="http://example.com:2908")}.}

\item{...}{passed to \code{\link{httr::POST}}.}
}
\value{
invisible the response of the server
}
\description{
Broadcast to a server via POST - hence this can run in another process or even on another host.
}
\examples{

\dontrun{
option(plotVF.default.broadcast=broadcastPost)
plotVR(iris)

# or directly:
plotVR(iris, broadcast=broadcastPost)
}
}
\seealso{
\cod{\link{plotVR}}
}

