% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caRdboard.R
\name{startBlockingServer}
\alias{startBlockingServer}
\title{Start the WebServer in a blocking version.}
\usage{
startBlockingServer(host = "0.0.0.0", port = 2908,
  interruptIntervalMs = ifelse(interactive(), 100, 1000))
}
\description{
Start the WebServer in a blocking version.
}
