% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvr.R
\name{plotVR}
\alias{plotVR}
\title{Plot data into the Cardboard VR}
\usage{
plotVR(data, col = NULL, broadcast = getOption("plotVF.default.broadcast",
  broadcastRefresh))
}
\description{
Plot data into the Cardboard VR
}

